services/price-ingestion-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
openpyxl==3.1.2
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1
redis==5.0.4
apscheduler==3.10.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
ruff==0.4.4
mypy==1.10.0
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
